Appt 565/okta integration (#454)
This change adds Okta authentication as an option for user login.


![mya-login](https://github.com/user-attachments/assets/6ab0cb05-c8a2-4c3f-8aa6-4956626edb94)

There is a test user set up in the okta organisation, credentials for
that will be available in confluence to facilitate testing of this
change, ~~however this will need actioning in whatever new development
and test organisations that will be set up. Details around the accounts
and organisation setup for the deployment to environments are still to
be determined.~~

NHS Mail should continue to work but as yet only been able to test with
the dummy oidc locally.

Significant change in this PR would be to the shape of the configuration
elements around auth. As there are now multiple auth providers that need
their own validation parameters and urls settting up I have altered the
configuration structure to facilitate this.

Obviously that will be a breaking change if the configuration for the
environments isn't amended before releasing this change set. The
intention would be to add this new configuration shape to the
environments in addition to the existing configuration before this is
released, and then the old configuration shape can be removed once this
release has stabilised.

---------

Co-authored-by: Vincent Crowe <vincent.crowe1@nhs.net>