APPT-634 Daily expired provisional cleanup (#527)
We don't need to be so aggressive with cleanup. NBS users can be on
the bookings page for an indeterminately long time, so might as well push the
cleanup to daily.

Tech Debt: Ideally would be nice to have a heartbeat functionality where
NBS slides our expiry time for a specific booking, if user is still on
the NBS page. No time for that though yet!!