APPT-588 - Bulk Import Endpoint (#516)
- A new bulk import endpoint that supports importing users / sites / API
users
- The new endpoint requires us to use Bearer authentication instead of
HMAC, so there is a new postman collection specifically for this new
endpoint as there is a pre-release script for the _Appointments Service_
postman collection which can't read the contents of the .csv files for
the new endpoint resulting in a signature mismatch and a 401 response
- There is some basic validation for both user and site csv files
themselves e.g. does the field have a value / is the value a valid
boolean
- There are further checks once we've mapped the contents of the csv
file.
- Sites - for siteIds we check to make sure they are distinct and that
the ODS Code is present in the well known ODS codes document store
- Users - I've added a check to make sure the siteId provided exists in
the site document store before we add the user
- A new permission required to access the endpoint -
`system:data-importer`

There are a couple of TODOs we need to iron out that are still in the
code and a couple of other things:
- Removal of the CsvDataTool from the solution in a separate PR. There
is some code duplication from there - the CsvProcessor was lifted from
it. Remove that later once we've made sure this solution has been tested
first.
- We need a test bearer token for this endpoint in postman against local
and dev environments - also applies for integration testing which isn't
included in this PR.
- Integration tests to be added once the above is sorted