APPT-627 - support querying availability by consecutive slots (#523)
This change enables NBS to query availability by consecutive slots. This
change allows NBS to pass 'consecutive' to the query availability
endpoint.

The changes doesn't overly effect the endpoint point past slots.
Anything above slots like daily will change the meaning of the count to
say the amount of consecutive availability there is. When query by slot
it will also return the consecutive slots but the logic is always based
on consecutive looking forward. A slot that returns 0 when query by
slots could be bookable because a slot before it has the calculation for
that slot, but the 0 slot has no consecutive slots when looking forward.