SRE-402: Enable joint bookings in staging (#629)
We need to performance test Joint Bookings. In order to do this, we need
to enable its feature flag in Staging.

For various reasons, this can currently only be done in Staging and
can't be done in Int. I don't like letting Staging deviate from what's
in Production, but frankly it already has and until we get a separate
perf environment this is the world we're in.