APPT-765 - Calculating the bookings from midnight of the day after the weekend (#626)
I've not update tests for this as Jonny is currently writing a
comprehensive test suite around timings in his branch and we agreed that
this change will be covered as part of it. On top of that, Saritha has a
PR for View Availability tests which cover this area as well.