APPT-749 Fix front end dateTime issue when the API is ran in a different timezone (#649)
Small missed edge case from me with the previous PR
https://github.com/NHSDigital/nbs-appointments-management-service/pull/638

When running the API in a different timezone also, I wanted to double
check that the comprehensive combination of node timezone and the
session tests still passed.

I noticed that the DailyAppointments page fails the tests in this case.
This is because it passes the datetime info into the booking query to
the API in UTC format, when it should be in the datetime timezone
unspecified format (secretly UK datetime).

Fundamentally, datetimes passed to the backend and parsed as a C#
DateTime should be in the EXACT same format as the database string
format: i.e "from": "2025-07-01T09:40:00".

We don't specify in the API that datetime objects are UK or do any
conversions, but so long as we always pass things around consistently it
should be okay. The backend code:
```
using (metricsRecorder.BeginScope("GetBookingsInDateRange"))
        {
            return await bookingStore.RunQueryAsync<Booking>(b => b.DocumentType == "booking" && b.Site == site && b.From >= from && b.From <= to);
        }
```
needs to have the DB datetime and the Query datetime in the same format
to work. I am not suggesting any changes to the backend for now 👍

Also fixed 'endOf(day)' as is more in line with a previous fix, and
removed the fetchAvailability method as it isnt used by frontend.

**No new tests need to be written**, as my previous PRs tests catch
this, its only when the API is in a different timezone does this become
an issue. I have reran all the combination tests I ran before (node
server in multiple different TZs, browser in multiple different TZs, and
NOW the API in multiple different TZs!)

PS - I imagine this COULD have eventually had an impact to LIVE MYA due
to UK/UTC differences, in the case of sessions/bookings being near
midnight (1 hour shift crossing day threshold), but I suspect we dont
have the data combination in it at the moment where this is an issue.