APPT-860: Create local stub of Okta directory (#662)
The secrets exposed here in Github are dev ones. We need to refresh
these and remove them from the codebase so they aren't exposed here any
longer.

We currently rely on hitting the dev instance of Okta for local
development and integration tests, so I've also created a local stub of
the Okta directly for use during integration tests.

For local development, we should override these values with
non-committed files.