APPT-704: Okta auth route (#661)
This PR creates codepaths for creating, editing and deleting users both
with and without the Okta feature flag. The key motivator for this work
was that the UI has changed substantially from its original design (new
designs in the [APPT-704
ticket](https://nhsd-jira.digital.nhs.uk/browse/APPT-704)), and
collecting the user's first/last names is now only required if the User
is A) an Okta user and B) does not yet exist in Okta.

Key changes: 
- The Okta code has been brought into the Core project because Core
needs to reference it directly
- There is a new API which, for a given email address, will inform on
whether that address already exists in MYA and its relevant Identity
Provider
- There is a new store for email domain whitelists which is sensitive to
the Okta feature toggle
- The new UI route has been created
- Unit, Integration and E2E tests for all the above

This already a MAHOOSIVE pull request, so there are some additional
changes I'll do as further tickets:
- I'd like much more substantial testing of these features with the
toggle on/off in Playwright. For now I've simply updated the existing
tests to match the new UI. I'd like to implement this using the tags
feature so we can selectively run tests in different modes, but that'll
trigger a fair raft of refactors to existing tests.
- The notification triggered after creating/editing a user is going to
be re-implemented in [its own
ticket](https://nhsd-jira.digital.nhs.uk/browse/APPT-799)

---------

Co-authored-by: ste-banks <stephen.banks4@nhs.net>