APPT-880 - Enabling the high load function for Bulk Import (#675)
We hit an issue on int where uploading 244 Okta users caused the
BulkImportFunction to hit it's default timeout (5 minutes) so for
staging / production we want to enable the high load function for this
which has a longer timeout (30 minutes as default).

Once we've figured out a limit on staging for high load, I will work
with Anna to workout a plan over a number of weeks to upload all users
(both nhs & Okta) so that we don't hit the timeout.