Add --strict to feature toggles import (#676)
There is no Jira ticket associated with this. The purpose of this PR is
to clean up the feature toggles (disable joint bookings) and to add the
--strict param to the import flags script. This new param means that it
will override all of the config in the app config. Meaning flags that
are not present in the imported json will be removed from app config.