APPT-885: Fix Okta provisioning bug (#682)
Attempting to add or edit an Okta user with 24 hours of first creating
them is causing an error.

This is because of a bug in some OktaService logic which assesses if
they're provisioned (i.e. invited to okta but not yet activated), and
re-activates that provisioning request if it's expired (i.e. they didn't
activate their account within 24 hours). This logic was returning an
unsuccessful result if they were provisioned but within that 24 hour
period, but it should've been returning a success.

This PR fixes this bug and also re-writes the service to be a bit
clearer/more readable.

---------

Co-authored-by: jsed-nhs <193929923+jsed-nhs@users.noreply.github.com>