APPT-814 - Introducing new regional user role (#696)
Introduction of a new regional user role `system:regional-user` which
grants a number of permissions to sites within the region
- Regional users **will not** appear in the manage users list if they
only have access to a site via the region.
- Regional users are able to manage users, view availability and manage
site details for sites within their region but they cannot create
availability (unless they're granted that permission on a site level).
- It is still possible for regional users to be granted site level
permissions to a site (e.g. Create availability).
- It is possible (at least for now) for users to be granted the regional
user permission to multiple regions - there may be further discussions
to restrict this to just one region.
- Small refactor of the PermissionChecker class to prevent multiple
fetches of permissions in each method. Now it will fetch user roles &
role assignments in parallel then filter on them instead. Confirmed all
existing unit tests still pass - which they do.
- Added new unit tests for the new region permissions.
- There are no e2e tests yet as there is a further ticket (APPT-868)
which involves front end changes for regional users so I figured it was
best to leave those for that ticket so we don't have to re-do them.
- Same applies for integration tests on the GetSitesPreviewFunction
endpoint - this will be changing as part of APPT-868 as well.