APPT-863: Validate coordinates in frontend. APPT-864: Allow empty phone number (#707)
We're not currently validating the Lat/Long co-ordinates when they're
edited through MYA itself.

This is part 1 of 2, which just adds our validation to the frontend.
Part 2 will be to add the same validation to the backend API.

Also doubles up as a PR for APPT-864, which is to permit the empty
string as a valid entry for the Phone Number field