APPT-970 - Removing the data-importer role and adding it as a permission to the admin role (#725)
Had a discussion with Lauren and Jack and opted to go down the route of
giving existing admin users the permission to import data via the bulk
upload rather than it being a specific role so this PR also remove the
`system:data-importer` role.

Integration tests for the bulk import are currently in a separate PR -
hence why they haven't been amended here.