APPT-938 - Adding bulk import integration tests (#710)
- Adding user & site bulk import integration tests using the feature
toggle pattern introduced by Jonny (hence the target branch).
- The tests cover the on & off states for the BulkImport endpoint for
both users and sites.
- These tests don't cover all the validation rules for the user & site
endpoints - just making sure these pass for both the on / off states
with valid csv files so we have confidence in enabling it
- Added a new base feature step which creates a site with a valid Guid
as the BulkImport will reject the file if it is not - have added a TODO
to tidy this up later. I was reluctant to make changes to the existing
step as it has a lot of uses