APPT-986: Fix erroneous notification config (#735)
We recently updated all `clinical_services.json` documents with the
confirmed services for the Autumn/Winter campaign.

What we (I) forgot to do, was update the notification config documents
with these new IDs. This means that when the NotifyFoo functions are
picking up messages from the bus, they are trying to fetch the
templateId for services which don't exist.

This PR fixes the documents and adds some tests to ensure that all
Notification Configurations use service Ids which exist in the clinical
services document.