APPT-986: Stub out notifications in Perf (#738)
We've ran into issues on the PERF environment, both when using the APIM
stub and a test profile of Gov.Notify. Because the notification
consumers run in a separate process, outside of the main
booking/querying journeys, they can be safely ignored without
invalidation the performance testing.

We could therefore just leave them erroring and not care, but the
problem then is that the dead-letter queue is growing to a huge size and
eventually maxing out. When it maxes out, the main BAU functions which
try to add events to that queue start erroring out too. So we need a way
to consume items from the queues without throwing any errors.

We already have a FakeNotificationsClient which isn't currently being
used. We also have the concept of a "Notifications Provider" which:

- When "none", will not consume notifications at all
- When "local", will write notifications to the console
- When "azure", will send notifications through to Gov.Notify as
required in prod

I have added a fourth, "azure-throttled", which utilises this found
FakeNotificationsClient.

I've also added in some tests to prove the correct service is being
registered.

This value only needs to be set in the PERF env, which isn't currently
in main. Therefore this code has been cherry-picked onto the perf branch
and the actual usage of this value is in [this
commit](https://github.com/NHSDigital/nbs-appointments-management-service/commit/10c9ccd28015fffc41be295a07c5866b2ec2b6e2)