Appt 861 - Bulk import regional users (#736)
- Amending the CSV to allow for region users with a new Region column
- Only one region scoped permission per user allowed per upload
- Validating the Region code is in the well known ODS document
- Exactly one of the site or region columns can be populated per row
- If site is populated, role assignment columns are mandatory
- If region is populated, then the role assignment columns are able to
be empty as the user will be assigned the `system:regional-ser` role
with the permissions that role grants
- Additional email address regex validation to ensure it is a valid
email address
- Cleaning up site import error messages so they're more clear to the
data importer
- Unit tests for all of the above

A separate ticket has been raised to address the removal of regional
users via the bulk import. This was split out as it will involve
touching code the front end also uses and this code doesn't currently
know the concept of a regional user.