APPT-956 & APPT-682 RFC Approval step for Production (#748)
**This change is a 2-birds-1-stone change - adding the RFC check in the
deploy step at the top will prevent the cases where we are having a
production lock state due to a timed out terraform plan step on
deployment to staging.**

Add EnterChangeRequest validation step in before any of the deploy
actions when the environment is Production.

This stops the terraform plan step from running (and acquiring a
Production state lock) when a release is deployed to Staging (since it
won't kick off the plan step without this manual validation passing).
Once this validation has timed out, the Production stage can be reran
(on the day of LIVE deployment) and if the validation passes with an
entered CFG, then it will do the plan and acquire the lock and hopefully
be successfully deployed and then release the lock.

This is a temporary workaround while we try and fine-tune our process.
This comment field can't be validated for valid Regex (or anything for
that matter!) and a workaround would have to be added for it.