APPT-1015: Close gap where Okta store is still used even when toggle is off (#765)
# Description

There's a bug in the UserService where we still try to check if a
potential new user already exists in Okta, even when the toggle is off.
The bug is low priority because we still correctly prevent that user
from being created, but in Production because the OktaUnimplementedStore
is being registered rather than a real service the frontend stalls.

Fixes # APPT-1015

# Checklist:

- [x] My work is behind a feature toggle (if appropriate)
- [x] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [x] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [x] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests