Appt 859/poc site report (#814)
# Description

This is an initial PR for Site Summary report. This not complete, but I
feel it's in a good spot to consider stopping making this PR too big.

This PR introduces - 

- Aggregated site capacity and booking data to enable reporting
- A CSV site Summary data report endpoint
- Ability to manually run aggregation

There's a lot more testing that needs to be applied, both unit and
integration. But I think we can apply this in future PR's, the advantage
we get from merging this now is that we get to test the POC (at this
stage) on Dev infra and make sure it works as expected on Azure Service
bus

Fixes # (issue)

# Checklist:

- [x] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [x] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [x] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [x] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests