APPT-1094: Reports Download page (#819)
# Description

Creates the Download Reports page described in [ticket
1094](https://nhsd-jira.digital.nhs.uk/browse/APPT-1094).

- Creates a new Datepicker component with tests
- Renames the dateFormat variable to RFC3339Format, because the
Datepicker strictly requires that we adhere to this format (which we are
anyway)
- Introduces a pattern of having to query for permissions at all sites
because the reporting button should appear on a site-specific page if
the permission exists at ANY site, even if not the current one. I'm
deeply against this pattern and echo @pata9 's calls to refactor the
permissions code as soon as practically possible.

@pata9 Looked into this and I've had to change it, the authentication
doesn't work here, so I've changed it to a normal fetch with the bearer
token.

- Write Playwright tests for the new UI
- Update the way the reports:sitesummary permission is obtained. As of
writing this PR it was in roles that are site scoped. We need to either
"shows this button when this user has this role at ANY site", or make a
new role that's global scoped. Either way, this needs updating.

@pata9 I've actioned this, added the ability to pass * as scope. Not
sure this is the right choice but I've created a permission ticket to
look into this more widely
https://nhsd-jira.digital.nhs.uk/browse/APPT-1167

- Question whether, if reports is indeed a globally scoped tool, there
should be a card linking to it on the Site page. I've put one in because
it's in the designs, but I don't think it makes sense given the
permissions context.

@pata9 trying to start discussions more widely on this

Fixes # (issue)

# Checklist:

- [x] My work is behind a feature toggle (if appropriate)
- [x] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [x] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [x] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [x] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [x] If I've made UI changes, I've added appropriate Playwright and
Jest tests

---------

Co-authored-by: pata9 <paul.tallet1@nhs.net>