APPT-1203 SitesByArea API SitesSupportService filtering option (#865)
# Description

Update to the GetSitesByAreaFunction to allow new optional filtering to
be used by NBS to restrict the long/lat area results to only return
sites that support the service being queried (this means that a site has
posted ANY availability within the daterange, it doesn't mean they have
capacity remaining - i.e we aren't doing full allocation and free slot
checks at this top level stage)

This is supposed to be 'temporary' in that we are going to attempt to
consolidate this API endpoint and the first api/query call into a new
single API endpoint in future (APPT-841)

**Having said that, the intention is for this to go out as a hotfix to
address A/W needs, so it needs to be performant and behave as
expected.**

Have added a LOT of integration tests to prove expected behaviours, as
well as old behaviour is still supported if needed to be rolled back to
by NBS.

Have added logging in for the 'Batching' of calls so that we can verify
in logs in Production whether the batch size needs increasing/decreasing
for certain services...?

The API does 2 calls, one to get all the cached sites in API and
calculated distance bounds in memory. Then I use all the potentially
valid sites (ordered distance ascending) to try and batch build up the
required results. If needed, it can iterate multiple times to find a
site that is the furthest away (whilst still within range) if its the
only one to support the service.

Fixes # (issue)

# Checklist:

- [x] My work is behind a feature toggle (if appropriate)
- [x] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [x] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [x] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [x] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [x] If I've made UI changes, I've added appropriate Playwright and
Jest tests

(cherry picked from commit 4e059ed2ea09113bb2956e41d4310ee0cc6b4809)