APPT-543: Improve error handling in NextJS server actions (#969)
# Description

The issue this PR addresses is the following: when a server action
throws an exception as part of an imperative action (e.g. form
submission, button click etc.) that exception is swallowed and the
frontend doesn't react to it at all.

- This is not the case during rendering. If an exception is thrown
during rendering, the closest error boundary is rendered instead.
- This is not the case if client-side exceptions are thrown. This also
correctly triggers the closest error boundary.
- [NextJS
docs](https://nextjs.org/docs/app/getting-started/error-handling)
recommend using the [useTransition
hook](https://react.dev/reference/react/useTransition) to wrap server
action invocations. Clien-side exceptions thrown within the usage of
this hook will trigger the nearest error boundary.
- The docs also recommend modelling errors as return values rather than
throwing exceptions. I'm disappointed in this, because I had hoped that
rather than having to code defensively and try/catch everything to
ensure we return an error model, we would be free to write code safe in
the knowledge that errors would invoke an error boundary.
- One solution to this would be to create an error component to render
in place of form components, but our requirements from UR/UX are that we
trigger the actual error page. We have no design for an error on a page
within the rest of the layout.

This PR: 

- Introduces the `ServerActionResult` type, which is either a success
model with the requested data or a failure model.
- Introduces a `fromServer` method to wrap server actions which receives
this model and either returns the requested data (so typescript types
remain the same and we don't have to guard against null/undefined
everywhere) or throws a client-side exception.
- Wraps all imperative invocations of server actions in the
`useTransition` hook, so any errors thrown by `fromServer` will trigger
the nearest error boundary.
- Updates the `appointmentService` to use chained promises and return
`Promise.reject` instead of throwing an error when it encounters
problems. This will have a minor performance impact.

Caveats:

- I really don't like having to make this change. I've tried really hard
to build the frontend in a way that avoids developers having to remember
special cases or explicitly handle errors at every opportunity. Things
like auth, layouts and error handling should come for free out of the
box. This change means that if you forget to wrap your data loading
methods in `fromServer()` you won't get the expected behaviour. I've
tried many different changes to avoid having to go down this route, and
talked the team through those on a call. Sadly this is the least-worst
solution I've come up with.

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests