APPT-875: Auto cancellations (#1024)
# Description

When NBS cancel bookings through their autocancellations process, this
needs to be recorded so it can be reported on.
To keep MYA generic and avoid implementing a CancellationReason which is
specific to NBS, we have agreed that:

- The CancellationReason enum in this case will be `CancelledByService`
- NBS will pass a property (the name/value of which is not yet decided)
in as `AdditionalData`.

Therefore this PR: 
* Creates a new CancellationReason called `CancelledByService`
* Adds an additionalData field onto the `CancelBookingRequest`. Any data
passed into this will patch (not override) any existing additionalData
on the booking.
* 
Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests