APPT-1252 - Removing the feature flag check on the cancel session endpoint (#1036)
# Description

This is a temporary change to remove the feature flag check on the
cancel session endpoint. The change was made but subsequent integration
tests which use that endpoint were not updated and it is blocking the
new edit session endpoint change from being merged. I will keep the
feature flag change in as it is needed elsewhere but there is already a
ticket to update all integration tests which use this endpoint to use
the new endpoint. Once that work has been done, this feature flag check
can be added back in.

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests