Appt 1352 - New edit session endpoint (#1028)
# Description

- A new endpoint designed to cover all aspects of session updates and
cancellations, it will also eventually replace the existing
_CancelSession_ endpoint when all the front end work is also completed.
- This endpoint is also protected by the new change session uplift
feature flag and will return a 501 if the feature flag is disabled.
- The payload will include a session matcher which can include session
properties (from, until, capacity, services, slot length) but it can
also be a wildcard ("*") which indicates a cancellation of all sessions
on a given date (or within a date range).
- The payload also includes a session replacement object which is the
same format as the session matcher (if updating a session) or it can be
null indicating that session (or sessions) are being cancelled.
- It will return a 422 response if the update or cancellation was
unsuccessful.
- Unit tests and happy path integration tests have been covered for all
of the above. I have raised a separate ticket (APPT-1401) which involves
re-purposing the _CancelSession_ integration tests to use this new
endpoint but as this PR is already big, I figured it should be separate.

The following scenarios are covered by this new endpoint:

- Cancel one session on one day
- Cancel all sessions on one day
- Cancel all sessions across multiple days
- Cancel one session on multiple days
- Edit one session on a one day
- Edit multiple sessions on multiple days

Fixes # (issue)

# Checklist:

- [x] My work is behind a feature toggle (if appropriate)
- [x] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [x] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [x] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests