APPT-653 - Removing the CSV data tool and associated code (#1070)
# Description

We have replaced the original bulk import tool with an API call and
we've been using it for months to onboard sites and users so the old one
is being removed in this PR. This includes:

- CsvDataTool project + files
- CsvDataTool unit tests
- Old bulkimport pipeline file
- Removal of any other references

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [x] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [x] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests